Made loginCredentials method in acrTools more efficient, added Pull Image from Azure option, temporary fix for no registries
