+++ explorer/deploy/webAppCreator.ts
@@ -9,9 +9,11 @@ import { ResourceManagementClient, ResourceModels, SubscriptionModels } from 'az
 import { Subscription } from 'azure-arm-resource/lib/subscription/models';
 import WebSiteManagementClient = require('azure-arm-website');
 import * as WebSiteModels from 'azure-arm-website/lib/models';
+import * as fs from 'fs';
+import * as path from 'path';
 import * as vscode from 'vscode';
 import { reporter } from '../../telemetry/telemetry';
-import { AzureImageNode } from '../models/azureRegistryNodes';
+import { AzureImageNode } from '../models/AzureRegistryNodes';
 import { DockerHubImageNode } from '../models/dockerHubNodes';
 import { AzureAccountWrapper } from './azureAccountWrapper';
 import * as util from './util';
