+++ explorer/models/azureRegistryNodes.ts
@@ -1,5 +1,5 @@
 import * as ContainerModels from 'azure-arm-containerregistry/lib/models';
-import { SubscriptionModels } from 'azure-arm-resource';
+import { ResourceManagementClient, SubscriptionClient, SubscriptionModels } from 'azure-arm-resource';
 import * as moment from 'moment';
 import * as path from 'path';
 import * as request from 'request-promise';
