+++ package.json
@@ -52,16 +52,22 @@
     "onCommand:vscode-docker.browseDockerHub",
     "onCommand:vscode-docker.browseAzurePortal",
     "onCommand:vscode-docker.explorer.refresh",
+<<<<<<< HEAD
     "onCommand:vscode-docker.delete-ACR-Registry",
     "onCommand:vscode-docker.delete-ACR-Repository",
     "onCommand:vscode-docker.delete-ACR-Image",
+=======
+    "onCommand:vscode-docker.deleteAzureImage",
+    "onCommand:vscode-docker.pullFromAzure",
+>>>>>>> Made loginCredentials method in acrTools more efficient, added Pull Image from Azure option, temporary fix for no registries
     "onView:dockerExplorer",
     "onDebugInitialConfigurations"
   ],
   "main": "./out/dockerExtension",
   "contributes": {
     "menus": {
-      "commandPalette": [{
+      "commandPalette": [
+        {
           "command": "vscode-docker.browseDockerHub",
           "when": "false"
         },
@@ -70,7 +76,8 @@
           "when": "false"
         }
       ],
-      "editor/context": [{
+      "editor/context": [
+        {
           "when": "editorLangId == dockerfile",
           "command": "vscode-docker.image.build",
           "group": "docker"
@@ -106,7 +113,8 @@
           "group": "docker"
         }
       ],
-      "explorer/context": [{
+      "explorer/context": [
+        {
           "when": "resourceFilename =~ /[dD]ocker[fF]ile/",
           "command": "vscode-docker.image.build",
           "group": "docker"
@@ -127,7 +135,8 @@
           "group": "docker"
         }
       ],
-      "view/title": [{
+      "view/title": [
+        {
           "command": "vscode-docker.explorer.refresh",
           "when": "view == dockerExplorer",
           "group": "navigation"
@@ -138,7 +147,8 @@
           "group": "navigation"
         }
       ],
-      "view/item/context": [{
+      "view/item/context": [
+        {
           "command": "vscode-docker.container.start",
           "when": "view == dockerExplorer && viewItem =~ /^(localImageNode|imagesRootNode)$/"
         },
@@ -195,8 +205,20 @@
           "when": "view == dockerExplorer && viewItem == dockerHubRootNode"
         },
         {
-          "command": "vscode-docker.delete-ACR-Repository",
-          "when": "view == dockerExplorer && viewItem == azureRepositoryNode"
+          "command": "vscode-docker.deleteAzureImage",
+          "when": "view == dockerExplorer && viewItem == azureImageNode"
+        },
+        {
+          "command": "vscode-docker.pullFromAzure",
+          "when": "view == dockerExplorer && viewItem == azureImageNode"
+        },
+        {
+          "command": "vscode-docker.browseDockerHub",
+          "when": "view == dockerExplorer && viewItem == dockerHubImageTag"
+        },
+        {
+          "command": "vscode-docker.browseDockerHub",
+          "when": "view == dockerExplorer && viewItem == dockerHubRepository"
         },
         {
           "command": "vscode-docker.delete-ACR-Image",
@@ -216,34 +238,40 @@
         }
       ]
     },
-    "debuggers": [{
-      "type": "docker",
-      "label": "Docker",
-      "configurationSnippets": [{
-        "label": "Docker: Attach to Node",
-        "description": "Docker: Attach to Node",
-        "body": {
-          "type": "node",
-          "request": "attach",
-          "name": "Docker: Attach to Node",
-          "port": 9229,
-          "address": "localhost",
-          "localRoot": "^\"\\${workspaceFolder}\"",
-          "remoteRoot": "/usr/src/app",
-          "protocol": "inspector"
-        }
-      }]
-    }],
-    "languages": [{
-      "id": "dockerfile",
-      "aliases": [
-        "Dockerfile"
-      ],
-      "filenamePatterns": [
-        "*.dockerfile",
-        "Dockerfile"
-      ]
-    }],
+    "debuggers": [
+      {
+        "type": "docker",
+        "label": "Docker",
+        "configurationSnippets": [
+          {
+            "label": "Docker: Attach to Node",
+            "description": "Docker: Attach to Node",
+            "body": {
+              "type": "node",
+              "request": "attach",
+              "name": "Docker: Attach to Node",
+              "port": 9229,
+              "address": "localhost",
+              "localRoot": "^\"\\${workspaceFolder}\"",
+              "remoteRoot": "/usr/src/app",
+              "protocol": "inspector"
+            }
+          }
+        ]
+      }
+    ],
+    "languages": [
+      {
+        "id": "dockerfile",
+        "aliases": [
+          "Dockerfile"
+        ],
+        "filenamePatterns": [
+          "*.dockerfile",
+          "Dockerfile"
+        ]
+      }
+    ],
     "configuration": {
       "type": "object",
       "title": "Docker configuration options",
@@ -403,7 +431,8 @@
         }
       }
     },
-    "commands": [{
+    "commands": [
+      {
         "command": "vscode-docker.configure",
         "title": "Add Docker files to Workspace",
         "description": "Add Dockerfile, docker-compose.yml files",
@@ -544,6 +573,11 @@
         "category": "Docker"
       },
       {
+        "command": "vscode-docker.pullFromAzure",
+        "title": "Pull Image from Azure",
+        "category": "Docker"
+      },
+      {
         "command": "vscode-docker.dockerHubLogout",
         "title": "Docker Hub Logout",
         "category": "Docker"
@@ -570,18 +604,22 @@
       }
     ],
     "views": {
-      "dockerView": [{
-        "id": "dockerExplorer",
-        "name": "Explorer",
-        "when": "config.docker.showExplorer == true"
-      }]
+      "dockerView": [
+        {
+          "id": "dockerExplorer",
+          "name": "Explorer",
+          "when": "config.docker.showExplorer == true"
+        }
+      ]
     },
     "viewsContainers": {
-      "activitybar": [{
-        "icon": "images/docker.svg",
-        "id": "dockerView",
-        "title": "Docker"
-      }]
+      "activitybar": [
+        {
+          "icon": "images/docker.svg",
+          "id": "dockerView",
+          "title": "Docker"
+        }
+      ]
     }
   },
   "engines": {
