+++ utils/Azure/acrTools.ts
@@ -5,9 +5,10 @@ import * as vscode from "vscode";
 import { NULL_GUID } from "../../constants";
 import { AzureImageNode, AzureRepositoryNode } from '../../explorer/models/AzureRegistryNodes';
 import { AzureAccount, AzureSession } from "../../typings/azure-account.api";
-import { AzureImage } from "../Azure/models/image";
-import { Repository } from "../Azure/models/Repository";
 import { AzureUtilityManager } from '../azureUtilityManager';
+import { AzureImage } from "./models/image";
+import { Repository } from "./models/repository";
+const teleCmdId: string = 'vscode-docker.deleteAzureImage';
 
 /**
  * Developers can use this to visualize and list repositories on a given Registry. This is not a command, just a developer tool.
@@ -214,27 +215,12 @@ export async function getAzureImages(element: Repository): Promise<AzureImage[]>
  * @param registry : the registry to get login credentials for
  * @param context : if command is invoked through a right click on an AzureRepositoryNode. This context has a password and username
  */
+///using for pull
 export async function loginCredentials(subscription: SubscriptionModels.Subscription, registry: Registry, context?: AzureImageNode | AzureRepositoryNode): Promise<{ password: string, username: string }> {
-    let node: AzureImageNode | AzureRepositoryNode;
-    if (context) {
-        node = context;
-    }
-    let username: string;
-    let password: string;
-
-    const client = AzureUtilityManager.getInstance().getContainerRegistryManagementClient(subscription);
-    const resourceGroup: string = registry.id.slice(registry.id.search('resourceGroups/') + 'resourceGroups/'.length, registry.id.search('/providers/'));
-
-    if (registry.adminUserEnabled) {
-        let creds = await client.registries.listCredentials(resourceGroup, registry.name);
-        password = creds.passwords[0].value;
-        username = creds.username;
-    } else {
-        //grab the access token to be used as a password, and a generic username
-        let creds = await getRegistryTokens(registry);
-        password = creds.accessToken;
-        username = NULL_GUID;
-    }
+    //grab the access token to be used as a password, and a generic username
+    let creds = await getRegistryTokens(registry);
+    let password = creds.accessToken;
+    let username = '00000000-0000-0000-0000-000000000000';
     return { password, username };
 }
